+++ ionoscloud/provider.go
@@ -93,6 +93,7 @@ func Provider() *schema.Provider {
 			"ionoscloud_k8s_cluster":          dataSourceK8sCluster(),
 			"ionoscloud_k8s_node_pool":        dataSourceK8sNodePool(),
 			"ionoscloud_backup_unit":          dataSourceBackupUnit(),
+			"ionoscloud_firewall":             dataSourceFirewall(),
 		},
 	}
 	provider.ConfigureContextFunc = func(ctx context.Context, d *schema.ResourceData) (interface{}, diag.Diagnostics) {